LPS-193258 Move BreadthFirstIterator and PostOrderIterator to another file class
