+++ modules/apps/object/object-api/src/main/java/com/liferay/object/definition/tree/Tree.java
@@ -5,16 +5,13 @@
 
 package com.liferay.object.definition.tree;
 
-import com.liferay.portal.kernel.util.ListUtil;
+import com.liferay.object.definition.tree.constants.TreeConstants;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Iterator;
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Objects;
-import java.util.Queue;
-import java.util.Stack;
 
 /**
  * @author Feliphe Marinho
@@ -66,15 +63,17 @@ public class Tree {
 	}
 
 	public Iterator<Node> iterator() {
-		return iterator("breadth-first");
+		return iterator(TreeConstants.ITERATOR_TYPE_BREADTH_FIRST);
 	}
 
 	public Iterator<Node> iterator(long objectDefinitionId) {
 		return new BreadthFirstIterator(getNode(objectDefinitionId));
 	}
 
-	public Iterator<Node> iterator(String iteratorStrategy) {
-		if (Objects.equals(iteratorStrategy, "breadth-first")) {
+	public Iterator<Node> iterator(String iteratorType) {
+		if (Objects.equals(
+				iteratorType, TreeConstants.ITERATOR_TYPE_BREADTH_FIRST)) {
+
 			return new BreadthFirstIterator(rootNode);
 		}
 
@@ -82,65 +81,5 @@ public class Tree {
 	}
 
 	protected final Node rootNode;
-
-	private static class BreadthFirstIterator implements Iterator<Node> {
-
-		public BreadthFirstIterator(Node node) {
-			_queue.add(node);
-		}
-
-		@Override
-		public boolean hasNext() {
-			return !_queue.isEmpty();
-		}
-
-		@Override
-		public Node next() {
-			Node node = _queue.poll();
-
-			List<Node> nodes = node.getChildNodes();
-
-			if (ListUtil.isNotEmpty(nodes)) {
-				_queue.addAll(nodes);
-			}
-
-			return node;
-		}
-
-		private final Queue<Node> _queue = new LinkedList<>();
-
-	}
-
-	private static class PostOrderIterator implements Iterator<Node> {
-
-		public PostOrderIterator(Node node) {
-			_fillStack(_stack.push(node));
-		}
-
-		@Override
-		public boolean hasNext() {
-			return !_stack.isEmpty();
-		}
-
-		@Override
-		public Node next() {
-			return _stack.pop();
-		}
-
-		private void _fillStack(Node node) {
-			List<Node> childNodes = node.getChildNodes();
-
-			if (childNodes == null) {
-				return;
-			}
-
-			for (int i = childNodes.size() - 1; i >= 0; i--) {
-				_fillStack(_stack.push(childNodes.get(i)));
-			}
-		}
-
-		private final Stack<Node> _stack = new Stack<>();
-
-	}
 
 }
